requirements.txt
bcrypt==4.1.2
black==24.2.0
blinker==1.7.0
click==8.1.7
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
httplib2==0.20.2
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llfuse==1.3.8
MarkupSafe==2.1.5
more-itertools==8.10.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.28
systemd-python==234
tomli==2.0.1
typing_extensions==4.10.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==3.0.1
zipp==1.0.0
